Massive improvement to UI, now dark themed Buttons are mostly converted to TextButton, in which a Stack of the button image and text is placed as a child, aligned in centre Background image is achieved by a Container decorated with the image below a transparent Scaffold AI still doesn't work Some fruits missing

+++ lib/settings_page.dart
@@ -2,7 +2,10 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:fruit_panda/settings_state.dart';
+import 'package:fruit_panda/utils.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+
+import 'constants.dart';
 
 class SettingsPage extends StatefulWidget {
   const SettingsPage({Key? key}) : super(key: key);
@@ -39,44 +42,47 @@ class _SettingsPageState extends State<SettingsPage> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text("Settings")),
-      body: Padding(
-        padding: const EdgeInsets.all(20),
-        child: Align(
-          alignment: Alignment.topLeft,
-          child: Column(
-            children: [
-              Row(
-                children: [
-                  const Expanded(child: Text("Music")),
-                  Checkbox(
-                    value: _state.music,
-                    onChanged: (val) => setState(() {
-                      _state.music = val!;
-                    }),
+    return WithBackground(
+      child: Scaffold(
+        appBar: AppBar(title: const Text("Settings")),
+        body: Padding(
+          padding: const EdgeInsets.all(20),
+          child: Align(
+            alignment: Alignment.topLeft,
+            child: Column(
+              children: [
+                Row(
+                  children: [
+                    const Expanded(child: Text("Music")),
+                    Checkbox(
+                      value: _state.music,
+                      onChanged: (val) => setState(() {
+                        _state.music = val!;
+                      }),
+                    ),
+                  ],
+                ),
+                Row(
+                  children: [
+                    const Expanded(child: Text("Sound")),
+                    Checkbox(
+                      value: _state.sound,
+                      onChanged: (val) => setState(() {
+                        _state.sound = val!;
+                      }),
+                    ),
+                  ],
+                ),
+                if (debug)
+                  ElevatedButton(
+                    onPressed: () async {
+                      final prefs = await SharedPreferences.getInstance();
+                      await prefs.clear();
+                    },
+                    child: const Text("Delete prefs"),
                   ),
-                ],
-              ),
-              Row(
-                children: [
-                  const Expanded(child: Text("Sound")),
-                  Checkbox(
-                    value: _state.sound,
-                    onChanged: (val) => setState(() {
-                      _state.sound = val!;
-                    }),
-                  ),
-                ],
-              ),
-              ElevatedButton(
-                onPressed: () async {
-                  final prefs = await SharedPreferences.getInstance();
-                  await prefs.clear();
-                },
-                child: const Text("Delete prefs"),
-              ),
-            ],
+              ],
+            ),
           ),
         ),
       ),
